+++ domain/user/src/main/kotlin/io/github/siyual_park/user/repository/UserRepository.kt
@@ -2,6 +2,7 @@ package io.github.siyual_park.user.repository
 
 import io.github.siyual_park.data.expansion.where
 import io.github.siyual_park.data.repository.r2dbc.R2DBCRepository
+import io.github.siyual_park.data.repository.r2dbc.SimpleR2DBCRepository
 import io.github.siyual_park.user.entity.User
 import org.springframework.dao.EmptyResultDataAccessException
 import org.springframework.data.r2dbc.core.R2dbcEntityOperations
@@ -10,9 +11,9 @@ import org.springframework.stereotype.Repository
 @Repository
 class UserRepository(
     entityOperations: R2dbcEntityOperations
-) : R2DBCRepository<User, Long>(
+) : R2DBCRepository<User, Long> by SimpleR2DBCRepository(
     entityOperations,
-    User::class,
+    User::class
 ) {
     suspend fun findByNameOrFail(name: String): User {
         return findByName(name) ?: throw EmptyResultDataAccessException(1)
