+++ module/auth/src/main/kotlin/io/github/siyual_park/auth/repository/ScopeTokenRepository.kt
@@ -3,6 +3,7 @@ package io.github.siyual_park.auth.repository
 import io.github.siyual_park.auth.entity.ScopeToken
 import io.github.siyual_park.data.expansion.where
 import io.github.siyual_park.data.repository.r2dbc.R2DBCRepository
+import io.github.siyual_park.data.repository.r2dbc.SimpleR2DBCRepository
 import kotlinx.coroutines.flow.Flow
 import org.springframework.dao.EmptyResultDataAccessException
 import org.springframework.data.r2dbc.core.R2dbcEntityOperations
@@ -11,7 +12,7 @@ import org.springframework.stereotype.Repository
 @Repository
 class ScopeTokenRepository(
     entityOperations: R2dbcEntityOperations
-) : R2DBCRepository<ScopeToken, Long>(
+) : R2DBCRepository<ScopeToken, Long> by SimpleR2DBCRepository(
     entityOperations,
     ScopeToken::class,
 ) {
